requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
logger-app==0.0.2.23091422
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
orjson==3.10.6
packaging==24.1
parso==0.8.4
passlib==1.7.4
pendulum==3.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
time-machine==2.15.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.4
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
